Make compatible with Java 1.7
This library uses Java 1.8 features, which unfortunately on Android would require an API version of 24 or above, or adding desugaring for Java 1.8 features. While it's easy to add desugaring for individual apps, this doesn't always play nicely with every app, and as such is not a recommended practice for libraries.

This change uses only Java 7 features, and copies over some Java 8 libraries from OpenJDK to assist with the transition (GPL-license unfortunately)